Refactored to avoid import dependencies
